# docs/swagger.yaml
openapi: 3.0.0
info:
  title: Medical Appointment API
  version: 1.0.0
  description: API documentation for Medical Appointment System
servers:
  - url: http://localhost:5000
    description: Development server
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    User:
      type: object
      required:
        - name
        - email
        - password
        - phoneNumber
        - CIN
        - role
      properties:
        name:
          type: string
          example: "Dr. John Smith"
        email:
          type: string
          format: email
          example: "john.smith@example.com"
        password:
          type: string
          format: password
          example: "Doctor123!"
        phoneNumber:
          type: string
          example: "+1234567890"
        CIN:
          type: string
          example: "D123456"
        role:
          type: string
          enum: [Doctor, Patient, Admin]
        specialization:
          type: string
          example: "Cardiology"
        department:
          type: string
          format: uuid
        profileImage:
          type: string
          format: uri
        diplomaImage:
          type: string
          format: uri
        availability:
          type: array
          items:
            type: object
            properties:
              day:
                type: string
                enum: [Monday, Tuesday, Wednesday, Thursday, Friday, Saturday, Sunday]
              slots:
                type: array
                items:
                  type: object
                  properties:
                    time:
                      type: string
                      example: "09:00"
                    isBooked:
                      type: boolean
                      default: false
        isValidated:
          type: boolean
          default: false
        medicalHistory:
          type: string
    Department:
      type: object
      required:
        - name
        - description
      properties:
        name:
          type: string
          example: "Cardiology"
        description:
          type: string
          example: "Diagnosis and treatment of heart diseases"
        doctors:
          type: array
          items:
            type: string
            format: uuid
    Appointment:
      type: object
      required:
        - patient
        - doctor
        - appointmentDate
        - reason
        - department
      properties:
        patient:
          type: string
          format: uuid
        doctor:
          type: string
          format: uuid
        appointmentDate:
          type: string
          format: date-time
        status:
          type: string
          enum: [Pending, Confirmed, Cancelled, Completed]
          default: Pending
        reason:
          type: string
        department:
          type: string
          format: uuid
    MedicalHistory:
      type: object
      required:
        - patient
        - doctor
        - diagnosis
      properties:
        patient:
          type: string
          format: uuid
        doctor:
          type: string
          format: uuid
        appointment:
          type: string
          format: uuid
        diagnosis:
          type: string
        prescription:
          type: array
          items:
            type: object
            properties:
              medication:
                type: string
              dosage:
                type: string
              frequency:
                type: string
              duration:
                type: string
        notes:
          type: string
        attachments:
          type: array
          items:
            type: object
            properties:
              type:
                type: string
              description:
                type: string
        vitalSigns:
          type: object
          properties:
            bloodPressure:
              type: string
            heartRate:
              type: string
            temperature:
              type: string
        followUpDate:
          type: string
          format: date-time
paths:
  /api/auth/register:
    post:
      tags: [Authentication]
      summary: Register new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/User'
      responses:
        201:
          description: User registered successfully
  /api/auth/login:
    post:
      tags: [Authentication]
      summary: Login user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - email
                - password
              properties:
                email:
                  type: string
                  format: email
                password:
                  type: string
                  format: password
      responses:
        200:
          description: Login successful
          content:
            application/json:
              schema:
                type: object
                properties:
                  token:
                    type: string
                  user:
                    $ref: '#/components/schemas/User'
...
  /api/doctors/{id}/appointments:
    get:
      tags: [Doctors]
      summary: Get doctor's appointments
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        200:
          description: List of doctor's appointments
...
  /api/departments:
    get:
      tags: [Departments]
      summary: Get all departments
      responses:
        200:
          description: List of all departments
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Department'
    
    post:
      tags: [Departments]
      summary: Create department
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Department'
      responses:
        201:
          description: Department created successfully
  /api/departments/{id}:
    get:
      tags: [Departments]
      summary: Get department by ID
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        200:
          description: Department details
    
    put:
      tags: [Departments]
      summary: Update department
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Department'
      responses:
        200:
          description: Department updated successfully
    
    delete:
      tags: [Departments]
      summary: Delete department
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        200:
          description: Department deleted successfully
  /api/appointments:
    get:
      tags: [Appointments]
      summary: Get all appointments
      security:
        - bearerAuth: []
      responses:
        200:
          description: List of all appointments
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Appointment'
    
    post:
      tags: [Appointments]
      summary: Create appointment
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Appointment'
      responses:
        201:
          description: Appointment created successfully
  /api/appointments/{id}:
    get:
      tags: [Appointments]
      summary: Get appointment by ID
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        200:
          description: Appointment details
    
    delete:
      tags: [Appointments]
      summary: Delete appointment
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        200:
          description: Appointment deleted successfully
  /api/appointments/{id}/status:
    put:
      tags: [Appointments]
      summary: Update appointment status
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                status:
                  type: string
                  enum: [Pending, Confirmed, Cancelled, Completed]
                medicalHistory:
                  $ref: '#/components/schemas/MedicalHistory'
      responses:
        200:
          description: Appointment status updated successfully
...
security:
  - bearerAuth: [] 